ProjetInfoL3/requirements.txt
# Requirements du projet

# Installer avec py -3 -m pip install -U -r requirements.txt
# ou pip3 install -U -r requirements.txt

# Paquets de base
pip
django
Pillow

# Extensions pour django
django-bootstrap4
django-hijack
django-hijack-admin

# Pour les tests
django-test-plus